Effects.MonadThread

Description

Provides the MonadThread typeclass.

Since: 0.1

Threads

class Monad m => MonadThread m where

Represents thread effects.

Since: 0.1

Methods

microsleep :: HasCallStack => Natural -> m ()

Runs sleep in the current thread for the specified number of microseconds.

Since: 0.1

sleep :: (HasCallStack, MonadThread m) => Natural -> m ()

Runs sleep in the current thread for the specified number of seconds.

Since: 0.1

Semaphores

class Monad m => MonadQSem m where

Effect for simple semaphores.

Since: 0.1

Methods

newQSem :: Int -> m QSem

Creates a QSem.

Since: 0.1

waitQSem :: QSem -> m ()

Waits for a QSem.

Since: 0.1

signalQSem :: QSem -> m ()

Signals a QSem.

Since: 0.1

newQSemN :: Int -> m QSemN

Creates a QSemN.

Since: 0.1

waitQSemN :: QSemN -> Int -> m ()

Waits for a QSemN.

Since: 0.1

signalQSemN :: QSemN -> Int -> m ()

Signals a QSemN.

Since: 0.1

Reexports

data Natural

Natural number

Invariant: numbers <= 0xffffffffffffffff use the NS constructor